@@ -1,14 +1,24 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+
 import { Provider } from 'react-redux';
 
 import App from './App';
+import reportWebVitals from './reportWebVitals';
+
 import store from './redux/store';
 
-const renderRoot = document.getElementById('root');
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <Provider store={store}>
-    <App />
-  </Provider>,
-  renderRoot
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </Provider>
 );
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
