@@ -1,6 +1,8 @@
-import React from "react";
 
-export const Header = () => {
+
+
+
+const Header = () => {
 
     return (
         <header>
