@@ -1,9 +1,9 @@
-import React from 'react';
 
-import { TABS } from '../redux/actions/types';
-import { toggleTab } from '../redux/actions';
+import { TABS } from "../redux/actions/type";
 
 import { useDispatch } from 'react-redux';
+
+import { toggleTab } from "../redux/actions";
 
 
 const Tabs = ({ currentTab }) => {
@@ -11,15 +11,14 @@ const Tabs = ({ currentTab }) => {
     const dispatch = useDispatch();
 
     return (
-        TABS.map(tab =>   
+        TABS.map(tab => (
             <button
-                key={tab}
                 className={tab === currentTab ? 'button selected' : 'button'}
                 onClick={() => dispatch(toggleTab(tab))}
             >
                 {tab}
             </button>
-        )
+        ))
     )
 }
 
