@@ -1,35 +1,37 @@
-import React, { useState } from 'react';
+import { useState } from "react";
 
-import { useDispatch } from 'react-redux'; 
-import { addNewTodo } from '../redux/actions';
+import { useDispatch } from "react-redux";
+
+import { addNewTodo } from "../redux/actions";
+
 
 const TodoForm = () => {
-  
-  const [text, setText] = useState('');
+    const [text, setText] = useState("");
 
-  const dispatch = useDispatch();
+    const dispatch = useDispatch();
 
-  const onInputChange = (e) => {
-    setText(e.target.value);
-  };
+    const onFormSubmit = (e) => {
+        e.preventDefault();
 
-  const onFormSubmit = (e) => {
-    e.preventDefault();
+        dispatch(addNewTodo(text));
 
-    dispatch(addNewTodo(text));
-    setText('');
-  };
+        setText('');
+    }
 
-  return (
-    <form className="form" onSubmit={onFormSubmit}>
-      <input
-        className="input"
-        value={text}
-        onChange={onInputChange}
-        placeholder="Enter new todo..."
-      />
-    </form>
-  );
+    const onInputChange = (e) => {
+        setText(e.target.value);
+    }
+
+    return (
+        <form className="form" onSubmit={onFormSubmit}>
+            <input  
+                placeholder="Enter new todo..."
+                className="input"
+                onChange={onInputChange}
+                value={text}
+            />
+        </form>
+    )
 }
 
 export default TodoForm;
