@@ -1,45 +1,61 @@
 import axios from 'axios';
-import {
-  GETALL_TODO,
-  ADDNEW_TODO,
-  TOGGLE_TODO,
-  UPDATE_TODO,
-  DELETE_TODO,
-  TOGGLE_TAB
-} from './types';
+
+import { ADDNEW_TODO, GETALL_TODO, TOGGLE_TODO, UPDATE_TODO, DELETE_TODO, TOGGLE_TAB } from './type';
 
 const API_URL = 'http://localhost:8000';
 
-export const getAllTodo = () => async (dispatch) => {
-  const res = await axios.get(`${API_URL}/todos`);
+export const addNewTodo = (data) => async (dispatch) => {
+    try {
+        const res = await axios.post(`${API_URL}/todos`, { data });
 
-  dispatch({ type: GETALL_TODO, payload: res.data });
-};
+        dispatch({ type: ADDNEW_TODO , payload: res.data });
+    } catch (error) {
+        console.log('Error while calling addNewTodo API ', error.message);
+    }
+}
 
-export const addNewTodo = (name) => async (dispatch) => {
-  const res = await axios.post(`${API_URL}/todos`, { name });
+export const getAllTodos = () => async (dispatch) => {
+    try {
+        const res = await axios.get(`${API_URL}/todos`);
 
-  dispatch({ type: ADDNEW_TODO, payload: res.data });
-};
+        dispatch({ type: GETALL_TODO , payload: res.data });
+    } catch (error) {
+        console.log('Error while calling getAllTodos API ', error.message);
+    }
+}
 
-export const toggleTodo = id => async dispatch => {
-  const res = await axios.post(`${API_URL}/todos/${id}`);
+export const toggleTodo = (id) => async (dispatch) => {
+    try {
+        const res = await axios.get(`${API_URL}/todos/${id}`);
 
-  dispatch({ type: TOGGLE_TODO, payload: res.data });
-};
+        dispatch({ type: TOGGLE_TODO , payload: res.data });
+    } catch (error) {
+        console.log('Error while calling getAllTodos API ', error.message);
+    }
+}
 
-export const updateTodo = (id, name) => async dispatch => {
-  const res = await axios.put(`${API_URL}/todos/${id}`, { name });
+export const updateTodo = (id, data) => async (dispatch) => {
+    try {
+        const res = await axios.put(`${API_URL}/todos/${id}`, { data });
 
-  dispatch({ type: UPDATE_TODO, payload: { ...res.data, name } });
-};
+        dispatch({ type: UPDATE_TODO , payload: res.data });
+    } catch (error) {
+        console.log('Error while calling updateTodo API ', error.message);
+    }
+}
 
-export const deleteTodo = id => async dispatch => {
-  const res = await axios.delete(`${API_URL}/todos/${id}`);
+export const deleteTodo = (id) => async (dispatch) => {
+    try {
+        const res = await axios.delete(`${API_URL}/todos/${id}`);
 
-  dispatch({ type: DELETE_TODO, payload: res.data });
-};
+        dispatch({ type: DELETE_TODO , payload: res.data });
+    } catch (error) {
+        console.log('Error while calling deleteTodo API ', error.message);
+    }
+}
 
 export const toggleTab = (tab) => async (dispatch) => {
-  dispatch({ type: TOGGLE_TAB, filter: tab });
-};
+    dispatch({ type: TOGGLE_TAB, selected: tab })
+}
+
+
