@@ -1,29 +1,25 @@
-import * as actionTypes from '../actions/types';
+import * as actionTypes from '../actions/type';
 
-export const todosReducer = (state = [], action) => {
-  switch (action.type) {
-    case actionTypes.GETALL_TODO:
-      return action.payload;
 
-    case actionTypes.ADDNEW_TODO:
-      return [action.payload, ...state];
+export const todosReducers = (state = [], action) => {
 
-    case actionTypes.TOGGLE_TODO:
-      return state.map(todo =>
-        todo._id === action.payload._id ? { ...todo, done: !todo.done } : todo
-      );
-
-    case actionTypes.UPDATE_TODO:
-      return state.map(todo =>
-        todo._id === action.payload._id
-          ? { ...todo, name: action.payload.name }
-          : todo
-      );
-
-    case actionTypes.DELETE_TODO:
-      return state.filter(todo => todo._id !== action.payload._id);
-      
-    default:
-      return state;
-  }
+    switch (action.type) {
+        case actionTypes.ADDNEW_TODO:
+            return [action.payload, ...state]
+        case actionTypes.GETALL_TODO:
+            return action.payload
+        case actionTypes.TOGGLE_TODO:
+            return state.map(todo => (
+                todo._id === action.payload._id ? { ...todo, done: !todo.done } : todo
+            ))
+        case actionTypes.UPDATE_TODO:
+            return state.map(todo => (
+                todo._id === action.payload._id ? { ...todo, data: action.payload.data } : todo
+            ))
+        case actionTypes.DELETE_TODO:
+            return state.filter(todo => todo._id !== action.payload._id);
+        
+        default: 
+            return state;
+    }
 }
