@@ -1,21 +1,21 @@
-import { createStore, applyMiddleware, combineReducers } from 'redux';
+import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
-import { todosReducer } from './reducers/todosReducer';
-import { currentTabReducer } from './reducers/currentTabReducer';
+import { todosReducers } from './reducers/todosReducer';
+import { tabReducer } from './reducers/tabReducer';
 
 const reducer = combineReducers({
-    todos: todosReducer, 
-    currentTab: currentTabReducer
+    todos: todosReducers,
+    currentTab: tabReducer
 })
 
 
 const middleware = [thunk];
 
 const store = createStore(
-    reducer, 
+    reducer,
     composeWithDevTools(applyMiddleware(...middleware))
-);
+)
 
 export default store;
