@@ -1,20 +1,22 @@
 import express from 'express';
 import cors from 'cors';
+import bodyParser from 'body-parser';
 
-// components
 import Connection from './database/db.js';
-import Routes from './routes/todos.js';
+import Routes from './routes/route.js';
 
-const PORT = 8000;
 const app = express();
 
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-app.use(cors())
 
+app.use(cors());
 
-app.use('/todos', Routes);
+app.use(bodyParser.json({ extended: true }));
+app.use(bodyParser.urlencoded({ extended: true }))
+
+app.use('/', Routes);
+
+const PORT = 8000;
 
 Connection();
 
-app.listen(PORT, () => console.log(`Server is running on PORT ${PORT}`));
+app.listen(PORT, () => console.log(`Your server is running successfully on PORT ${PORT}`));
